+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 // import { useState } from 'react';
 // import PropTypes from 'prop-types';
-import Calculator from './components/calculator';
+import Calculator from './components/Calculator';
 
 function App() {
   return (
